Phase 3: per-phase resistance from actual wire length + EM symbol checks
- Add phase_resistance.py: R=ρL/A from actual phase path lengths,
  using BARE copper cross-section (not insulated diameter)
- Per-phase resistances differ (U<V<W) because path lengths differ;
  no hidden correction constant to force balance
- Temperature correction with explicit α; joint resistance待确认
- EM symbol verification: independent coil reference (shared topology),
  three-phase phasor balance, 120° sequence, rotation reversal
- Verification chain documents: geometryA→pathA→mesh NOT connected→
  flux NOT computed→torque NOT computed; old model not reused
- simulation_ready=false, old_model_fallback=false
- 8 new tests; 52 total pass

+++ organic_motor/construct/wiring_bracket.py
@@ -172,9 +172,9 @@ def build_clamp_assets():
         clamps[name] = cable_clamp(z_level, r_bridge, angle)
         clamp_list.append(dict(
             name=name, phase=phase_label, bridge_idx=bridge_idx,
-            r_mm=r_bridge, z_mm=z_level, angle_rad=float(angle),
-            position_mm=[r_bridge * np.cos(angle),
-                         r_bridge * np.sin(angle), z_level]))
+            r_mm=float(r_bridge), z_mm=float(z_level), angle_rad=float(angle),
+            position_mm=[float(r_bridge * np.cos(angle)),
+                         float(r_bridge * np.sin(angle)), float(z_level)]))
     return clamps, clamp_list
